// src/components/home/types.ts
export type Vertical = {
    _id: string;
    title?: string;
    shortDescription?: string;
    verticalImage?: string;
    knowMoreUrl?: string;
};

export type Recognition = {
    _id: string;
    recognitionImage?: string;
    recognitionName?: string;
    awardYear?: string;
};

export type Differentiator = {
    _id: string;
    title?: string;
    description?: string;
};